12_OpenAI_Agents_SDK/02_hello_agent/without_ui/pyproject.toml
[project]
name = "without-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.3.0",
]
